backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
google-adk>=0.1.0
ag-ui-adk>=0.3.4
google-api-python-client>=2.100.0
httpx>=0.25.0

